requirements.txt
flask
flask_cors
flask_session
#bson
pymongo
bcrypt
psycopg2
pandas
pyodbc
msal
oracledb
cx_Oracle
python-dotenv